:recycle: Modularizado
